fixed weight goal display and changed social media display
The weight goal display has been slightly modified for people who have a goal of losing weight, social media sharing will share weight goal value as well, social media sharing only available when weight goal is hit
