+++ frontend/src/components/landing-page/Landing.vue
@@ -18,17 +18,12 @@
 
         <li class="nav-item">
           <div class = "top-buffer"></div>
-            <a class="nav-link" href="/dashboard"><span class="mb-1 h4">See Progress</span></a>
-        </li>
-
-        <li class="nav-item">
-          <div class = "top-buffer"></div>
             <a class="nav-link" href="/edit-profile"><span class="mb-1 h4">Edit Profile</span></a>
         </li>
 
         
           <div class = "top-buffer"></div>
-            <shareNetworks />
+          <shareNetworks :weightPercentage="weightPercentage" :goalWeight = "goalWeight" />
       
 
         <li class="nav-item">
@@ -54,8 +49,19 @@
 
     </div>
   
-      <!-- Weight progress bar -->
-       <currentWeight/>
+          <div class="container">
+          <div class="panel panel-default">
+            <template v-if="weightPercentage == 100">
+            <p>Congrats! you hit your goal of {{goalWeight}}lb!</p>
+            </template>
+
+            <template v-else>
+              <div class="class=flex-container"><p>Current progress : {{currentWeight}}lbs/{{goalWeight}}lbs</p></div>
+              <p>Keep it up, you're {{ weightPercentage }}% towards completing your goal!</p>
+            </template>
+                </div>
+            
+        </div>
       <div class = "top-buffer"></div>
     
   
@@ -173,7 +179,6 @@ import todayMacros from './TodayMacros.vue';
 import todayStep from './TodaySteps.vue';
 import todayCalBurned from './TodaysCalorieBurn.vue';
 import shareNetworks from './ShareNet.vue';
-import currentWeight from './WeightGoals.vue';
 
 
 export default{
@@ -189,7 +194,6 @@ export default{
       SleepTrackChart,
       moodChart,
       shareNetworks,
-      currentWeight
     },
 
     
@@ -199,6 +203,10 @@ export default{
         workoutlist : [],
         averageSleep: 0, // Add this property to store the average sleep data
         currentName : '',
+        weightPercentage: 0,
+        currentWeight : 0,
+        goalWeight : 0,
+
     
 
       };
@@ -224,14 +232,58 @@ export default{
 
       axios.get(`http://127.0.0.1:5000/get_exercise/${currentuser}`).then(response => this.workoutlist = response.data)
       
-      
-     
+      let currentUserData = await axios.get(`http://127.0.0.1:5000/user_api/get_user_data/${currentuser}`).catch(function(error){
+            console.log(error);
+        });
+        let currentGoals = await axios.get(`http://127.0.0.1:5000/goals_api/get_user_goals/${currentuser}`).catch(function(error){
+            console.log(error);
+          });
+
+        //Getting current goals for today
+        let goalsList = currentGoals.data["user_goals"];
+        var today = new Date();
+        let todayFormatted = today.toLocaleDateString();
+        for(let i in goalsList){
+            let packet = goalsList[i];
+            let current = new Date(packet[4]);
+            let currentCompare = (current.getUTCMonth() + 1) + "/" + current.getUTCDate() + "/" + current.getUTCFullYear();
+            if(currentCompare == todayFormatted){
+                this.goalWeight = packet[1];
+                continue;
+            }
+         }
+
+        //Getting current weight of today
+        let weightList = currentUserData["data"]["user_data"][0][2];
+        this.currentWeight = weightList
+        
+
+
+        if(this.currentWeight<=this.goalWeight){
+          let progressCalc = Math.round((this.currentWeight/this.goalWeight)*100);
+          this.weightPercentage = progressCalc
+          localStorage.setItem("goal",this.weightPercentage);
+        }
+        else if(this.currentWeight>this.goalWeight){
+          if(this.goalWeight == 0){
+            this.weightPercentage = 0;
+            localStorage.setItem("goal",this.weightPercentage);
+          }
+          else{
+          let progressCalc = Math.round(((this.currentWeight-this.goalWeight)/this.currentWeight)*100);
+          this.weightPercentage = 100 - progressCalc;
+          localStorage.setItem("goal",this.weightPercentage);
+          }
+          
+        }
+     console.log(localStorage.getItem('goal'))
 
     },
     
 
 
     create(){
+      this.weightPercentage = localStorage.getItem("goal");
       this.mounted();
     },
